requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.2
sqlparse==0.4.4
tzdata==2024.1

# Dependencias adicionales necesarias
Pillow>=10.0.0
djoser==2.2.2
python-decouple==3.8
